[#AIO-964] Addresses PR comments
